tests: reuse vsock helper methods
